piecharts

+++ Common/ZDB/Models/MongoCollectionModel.cs
@@ -3,9 +3,9 @@
     public class MongoCollectionModel
     {
         public string Name { get; set; }
-        public int Count { get; set; }
-        public int Size { get; set; }
-        public int StorageSize { get; set; }
+        public double Count { get; set; }
+        public double Size { get; set; }
+        public double StorageSize { get; set; }
 
         public MongoCollectionModel()
         {
